refactor: change user to auth for better meaning, hash password♻️
