+++ apps/api/package.json
@@ -20,6 +20,7 @@
     "@hono/node-server": "^1.12.2",
     "@hono/zod-validator": "^0.2.2",
     "@paralleldrive/cuid2": "^2.2.2",
+    "bcrypt": "^5.1.1",
     "better-sqlite3": "^11.2.1",
     "cross-env": "^7.0.3",
     "dotenv": "^16.4.5",
@@ -29,6 +30,7 @@
     "zod": "^3.23.8"
   },
   "devDependencies": {
+    "@types/bcrypt": "^5.0.2",
     "@types/better-sqlite3": "^7.6.11",
     "@types/node": "^20.11.17",
     "drizzle-kit": "^0.24.2",
