+++ utils/api-seatsavvy/users/Create User.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 post {
-  url: http://localhost:5500/api/v1/users
+  url: http://localhost:5500/api/v1/auth/register
   body: json
   auth: none
 }
